refact: 결제 생성 API를 payment 도메인으로 이동 #36
